% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{river the fish was caught in}
  \item{MILE}{distance in that the fish was caught}
  \item{SPECIES}{species of fish}
  \item{LENGTH}{length of fish}
  \item{WEIGHT}{weight of fish}
  \item{DDT}{ddt value associated with each fish}
  ...
}
}
\source{
<https://books.google.com/books/about/Statistics_for_Engineering_and_the_Scien.html?id=Fdi_DQAAQBAJ>
}
\usage{
ddt
}
\description{
A data set of 144 fish that were caught in different rivers
}
\keyword{datasets}
