% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmytable.R
\name{ggmytable}
\alias{ggmytable}
\title{ggmytable}
\usage{
ggmytable(df, x, y, z)
}
\arguments{
\item{df}{The desired data frame to be analyzed}

\item{x}{A qualitative variable designated to the x axis}

\item{y}{A qualitative variable to group the data}

\item{z}{A quantitative variable used to calculate mean within the summary}
}
\value{
A bar graph of x, facet wrapped with y, and a summary of function n along with the mean of z
}
\description{
ggmytable
}
\examples{
\dontrun{ggmytable(mtbe, "MTBE-Detect", "Aquifier", "MTBE-Level")}
}
